+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - add: notificationForceTimeout option
 - add: bluetooth device-added, device-removed signal
 - add: cursor property
+- feat: window popup close on click away
 
 ## Breaking Changes